Декомпозиция UserController.js
