+++ api/controllers/UserController/logout.js
@@ -1,0 +1,7 @@
+
+module.exports = function logout(req, res) {
+  
+  delete req.session.user;
+  return res.redirect('/');
+  
+};
