+++ api/controllers/UserController/register.md
@@ -1,0 +1,151 @@
+
+Диаграммы последовательностей для действия `UserController.register()`
+======================================================================
+
+
+1. Когда ошибок не происходит (позитивный сценарий):
+```
+Browser                           UserController  Waterline  nodemailer
+  |                                 |               |          |
+  |   POST /register?username=...   |               |          |
+  +-------------------------------->|               |          |
+  |                                 |
+  |                                 +--+ Разбор параметров запроса:
+  |                                 |  | model = req.allParams();
+  |                                 |<-+ 
+  |                                 |
+  |                                 | User.create() |          |
+  |                                 +-------------->|          |
+  |                                 |     data      |          |
+  |                                 |<--------------+          |
+  |                                 |                          |
+  |                                 |  sendMail(model.email)   |
+  |                                 +------------------------->|
+  |                                 |           info           |
+  |                                 |<-------------------------+
+  |                                 |
+  |                                 +--+ Запоминаем почтовый
+  |                                 |  | адрес в сессии:
+  |                                 |<-+ session.flash = {model.email}
+  | redirect(303, '/after_register' |
+  |          "подтвердите почту")   |               |          |
+  |<--------------------------------+               |          |
+  |
+  |  Представление user/after_register, кроме сообщения о необходимости
+  |  подтвердить почтовый адрес, может показать ссылку на почтовую систему,
+  |  которую использует пользователь. Передача значения почтового адреса
+  |  в представление выполняется с помощью middleware policies/flash.js
+  |  через сессию:
+  |
+  |                           sails
+  | GET /after_register         |
+  +---------------------------->|    policies/flash.js
+  |                             |      |
+  |                             +----->|
+  |                             |      +--+
+  |                             |      |  | res.locals.flash = req.session.flash
+  |                             |      |<-+ 
+  |                             |<-----+
+  |                             |
+  | res.view('/after_register') |
+  |<----------------------------+
+
+  Затем пользователь открывает свой почтовый ящик
+  и переходит по предложенной в письме ссылке:
+  
+  |   GET /register/?id=...&t=...   |               |          |
+  +-------------------------------->|               |          |
+  |                                 | User.findOne()|          |
+  |                                 +-------------->|          |
+  |                                 |     user      |          |
+  |                                 |<--------------+          |
+  |                                 |
+  |                                 +--+
+  |                                 |  | Отмечаем, что пользователь
+  |                                 |  | теперь может входить на сайт:
+  |                                 |  | user.active = true;
+  |                                 |<-+
+  |                                 |
+  |                                 | User.update() |          |
+  |                                 +-------------->|          |
+  |                                 |<--------------+          |
+  |     redirect(303, '/login')     |               |          |
+  |<--------------------------------+               |          |
+```
+
+2. Если произошла ошибка при обработке запроса `GET /register/?id=...&t=...`.
+Например, при работе с базой данных (поиске, либо обновлении записи о пользователе):
+```
+  |                                 |  error  
+  |                                 |<-------O
+  |  res.view('/user/error',        |               |          |
+  |           {'описание ошибки'})  |               |          |
+  |<--------------------------------+               |          |
+```
+
+3. Если произошла ошибка при обработке запроса `POST /register`. Например, 
+при разборе параметров запроса или при работе с базой данных
+(при создании, либо удалении записи о пользователе):
+```
+  |                                 |  error  
+  |                                 |<-------O
+  |                                 |
+  |                                 +--+ Запоминаем ошибку и заполненные поля:
+  |                                 |  | session.flash = {'описание ошибки',
+  |                                 |<-+                  model}
+  |    redirect(303, '/register')   |
+  |<--------------------------------+               |          |
+  |                                 |               |          |
+  |       GET /register             |               |          |
+  +-------------------------------->|               |          |
+  |                                 |               |          |
+  |    res.view('/register',        |               |          |
+  |             {session.flash})    |               |          |
+  |<--------------------------------+               |          |
+```
+
+4. Если произошла ошибка при отправке почтового сообщения (пример ошибки,
+возникающей при обработке другой ошибки в обработчике `POST /register`):
+```
+  |                                 |          error           |
+  |                                 |<-------------------------+
+  |                                 |
+  |                                 +--+ Запоминаем ошибку и заполненные поля:
+  |                                 |  | session.flash = {'описание ошибки',
+  |                                 |<-+                  model}
+  |                                 |
+  |                                 | Удаляем только-что созданную запись:
+  |                                 | User.delete()
+  |                                 +-------------->|
+  |                                 |<--------------+
+  |                                 | Если при удалении записи произойдет
+  |                                 | ошибка, то переходим к варианту 2).
+  |    redirect(303,'/register')    |
+  |<--------------------------------+               |          |
+  |                                 |               |          |
+  |       GET /register             |               |          |
+  +-------------------------------->|               |          |
+  |                                 |               |          |
+  |    res.view('/register',        |               |          |
+  |             {session.flash})    |               |          |
+  |<--------------------------------+               |          |
+```
+
+Примечания:
+
+После получения запроса POST сервер должен всегда выполнять редирект
+с кодом 303.
+
+Данные внутренних ошибок (Error().message) должны выводиться только
+в серверной консоли. В браузер передается только общее описание ошибки.
+
+Представление user/register должно заполнять поля и текст ошибки
+из session.flash, если есть.
+
+Представление user/after_register должно использовать почтовый адрес
+указанный в session.flash (если есть) для того, чтобы по возможности
+предоставить ссылку на почтовую систему, которую использует пользователь.
+
+Политика доступа к профилю пользователя (policies/sessionAuth.js) должна
+проверять флаг user.active.
+
